restructure for dockerization
