+++ Dockerfile
@@ -10,19 +10,13 @@ RUN apk update && apk upgrade && \
 WORKDIR /app
 
 # Copy go mod and sum files
-COPY go.mod go.sum ./
+COPY mixerlib/go.mod go.sum ./
 
 # Download all dependencies. Dependencies will be cached if the go.mod and go.sum files are not changed
 RUN go mod download
 
 # Copy the source from the current directory to the Working Directory inside the container
-COPY . .
-
-# Build the Go app
-RUN go build -o main ./cmd/mixer-cli
-
-# Expose port 8080 to the outside world
-EXPOSE 8080
+COPY .. .
 
 # Run the executable
-CMD ["./main"]
+CMD [ "go", "run", "./cmd/mixer-cli/main.go" ]
