+++ clientlib/go.mod
@@ -1,0 +1,8 @@
+module client
+
+go 1.16
+
+require (
+	github.com/google/uuid v1.2.0
+	github.com/oliviabarnett/mixer v0.0.0-20210418181659-75f20f9cc0d6
+)
