+++ docker-compose.yml
@@ -1,21 +1,14 @@
-# Docker Compose file Reference (https://docs.docker.com/compose/compose-file/)
-
 version: '3'
 
-# Define services
 services:
-
-  # App Service
-  mixer:
-    build:
-      context: . # Use an image built from the specified dockerfile in the current directory.
-      dockerfile: Dockerfile
+  clientmodule:
+    image: clientmodule
+    build: .
+  mixermodule:
+    image: mixermodule
+    build: .
+    # ports exposed to localhost
     ports:
-      - "8080:8080" # Forward the exposed port 8080 on the container to port 8080 on the host machine
-    restart: unless-stopped
-    networks: # Networks to join (Services on the same network can communicate with each other using their name)
-      - backend
-
-# Networks to be created to facilitate communication between containers
-networks:
-  backend:    
+      - 7070:7070
+    depends_on:
+      - clientmodule
