+++ pkg/dispatcher.go
@@ -1,4 +1,4 @@
-package internal
+package pkg
 
 import (
 	"fmt"
