+++ pkg/dispatcher_test.go
@@ -1,4 +1,4 @@
-package internal
+package pkg
 
 import (
 	"fmt"
