+++ pkg/transaction.go
@@ -1,4 +1,4 @@
-package internal
+package pkg
 
 import (
 	"crypto/sha256"
